fix: 検索キャンセルの処理が行えるように実装
